requirements.txt
flask
faiss-cpu
open_clip_torch
torch
torchvision
Pillow
firebase-admin
google-generativeai
algoliasearch
